feat: Remove unused validate command from CLI
The `validate` command in the CLI has been removed as it is no longer needed. This command was responsible for validating the configuration file, but this functionality has been moved to the main application logic. By removing the `validate` command, we simplify the CLI and focus on the core functionality of the application.